% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_script.R
\name{create_script}
\alias{create_script}
\title{Create a Script with Header}
\usage{
create_script(filename, title = NULL)
}
\arguments{
\item{filename}{string (without .R)}

\item{title}{string}
}
\description{
Creates a scripts with documentation header without the need to copy and paste things into the editor.
}
\examples{
create_script("my_new_script", "My New Script")
}
