% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdPrime.r
\name{gcdPrime}
\alias{gcdPrime}
\title{Greatest Common Divisor}
\usage{
gcdPrime(num1, num2)
}
\arguments{
\item{num1}{The first integer}

\item{num2}{The second integer}
}
\value{
gcd
}
\description{
This function computes the greatest common divisior (GCD) of two positive integers using the Euclidian algorithm.
}
\details{
This was written some abstract algebra class.
}
\examples{
gcdPrime(16,24)
}
