% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_converter.R
\name{ggplot_converter}
\alias{ggplot_converter}
\title{ggplot2 converter function}
\usage{
ggplot_converter(p, want.bw = TRUE, axis.text.size = 15,
  axis.title.size = 15, plot.facet = FALSE, strip.text.size = 15,
  main.title = FALSE, plot.title.size = 15, have.legend = FALSE,
  legend.text.size = 10, legend.title.size = 10)
}
\arguments{
\item{p}{A ggplot2 object}

\item{want.bw}{Change background to white}

\item{axis.text.size}{Axis text size}

\item{axis.title.size}{Axis title size}

\item{plot.facet}{Plot figures across a factor}

\item{main.title}{logical. Main title}

\item{plot.title.size}{Main title size}

\item{have.legend}{logical. Legend present in the figure}

\item{legend.text.size}{Legend text size}

\item{legend.title.size}{Legend title size}
}
\description{
This function will take a ggplot object p and convert it into a plot that is suitable for a journal article.
}
\details{
This function is based on some code I got from Brian Weaver.
}
\examples{


}
\author{
Brian Weaver, Isaac Michaud
}
