% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scale.R
\name{logseq}
\alias{logseq}
\title{Logarithmic sequence}
\usage{
logseq(from = 1, to = 1, length, want.base.e = TRUE)
}
\arguments{
\item{from}{lower endpoint}

\item{to}{upper endpoint}

\item{length}{number of points}

\item{want.base.e}{logical. If FALSE then used base-10 logarithm}
}
\description{
Logarithmic sequence
}
