% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgroups.R
\name{plotgroups.barplot}
\alias{plotgroups.barplot}
\title{Plot several groups of repeated observations as a barplot.}
\usage{
plotgroups.barplot(data, stats, colors, ylim, features, barwidth,
  whiskerswidth = barwidth, whiskerslwd = par("lwd"),
  whiskerscol = "black", bordercol = "black", ...)
}
\arguments{
\item{data}{list, each element is a vector of replicates for one combination of parameters}

\item{stats}{list containing the summary statistics (see \code{\link{plotgroups}})}

\item{colors}{colors for plotting}

\item{ylim}{whether to compute Y axis limits}

\item{features}{the features to plot. At the moment, either \code{mean} or \code{median} can be plotted. Also, only one of \code{box}, \code{iqr}, \code{sd}, \code{sem} can be plotted at the moment.}

\item{barwidth}{width of the individual bars/boxes etc. as fraction of 1}

\item{whiskerswidth}{width of the whiskers as fraction of 1}

\item{whiskerslwd}{line width of the whiskers}

\item{whiskerscol}{color of the whiskers}

\item{bordercol}{color of the border}

\item{...}{additional parameters passed to \code{\link[graphics]{rect}}}
}
\description{
This is intended to be called from \code{\link{plotgroups}}.
}
\seealso{
\code{\link{plotgroups}}

Other plotgroups.plotting.functions: \code{\link{plotgroups.beeswarm}},
  \code{\link{plotgroups.boxplot}},
  \code{\link{plotgroups.vioplot}}
}

