% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgroups.R
\name{plotgroups.beeswarm}
\alias{plotgroups.beeswarm}
\title{Plot several groups of repeated observations as a beeswarm plot.}
\usage{
plotgroups.beeswarm(data, stats, colors, ylim, features, barwidth, palpha = 1,
  bxplwd = par("lwd"), bxpcols = colors, ...)
}
\arguments{
\item{data}{list, each element is a vector of replicates for one combination of parameters}

\item{stats}{list containing the summary statistics (see \code{\link{plotgroups}})}

\item{colors}{colors for plotting}

\item{ylim}{whether to compute Y axis limits}

\item{features}{the features to plot. At the moment, either \code{mean} or \code{median} can be plotted. Also, only one of \code{box}, \code{iqr}, \code{sd}, \code{sem}, \code{ci} can be plotted at the moment.}

\item{barwidth}{width of the individual bars/boxes etc. as fraction of 1}

\item{palpha}{opacity of the individual points}

\item{bxplwd}{line width for the simplified boxplot}

\item{bxpcols}{colors for the simplified boxplot}

\item{...}{additional parameters passed to \code{\link[beeswarm]{beeswarm}}}
}
\value{
Same as \code{\link[beeswarm]{beeswarm}}
}
\description{
This is intended to be called from \code{\link{plotgroups}}. Requires the \code{beeswarm} package.
}
\seealso{
\code{\link{plotgroups}}

\code{\link[beeswarm]{beeswarm}}

Other plotgroups.plotting.functions: \code{\link{plotgroups.barplot}},
  \code{\link{plotgroups.boxplot}},
  \code{\link{plotgroups.vioplot}}
}

