% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgroups.R
\name{plotgroups.boxplot}
\alias{plotgroups.boxplot}
\title{Plot several groups of repeated observations as a boxplot (extended with mean, standard deviation, and standard error of the mean).}
\usage{
plotgroups.boxplot(data, stats, colors, ylim, features, barwidth, bxppars, ...)
}
\arguments{
\item{data}{list, each element is a vector of replicates for one combination of parameters}

\item{stats}{list containing the summary statistics (see \code{\link{plotgroups}})}

\item{colors}{colors for plotting}

\item{ylim}{whether to compute Y axis limits}

\item{features}{the features to plot.}

\item{barwidth}{width of the individual bars/boxes etc. as fraction of 1}

\item{bxppars}{additional parameters passed to \code{\link[graphics]{boxplot}} as
sig\code{pars} parameter}

\item{...}{additional parameters passed to \code{\link[graphics]{boxplot}} as \code{...}. Can also
contain the following:
\describe{
          \item{meanlty, meanlwd, meancol, meanpch, meancex}{Mean line type, line width,
                color, point character, and point size expansion. The default
                \code{meanpch=NA} suppresses the point, and \code{meanlty="blank"}
                does so for the line. Note that \code{meanlwd} defaults to 3x the
                default lwd and \code{meancol} defaults to \code{"red"}}
          \item{sdwhisklty, sdwhisklwd, sdwhiskcol}{Whisker line type(default:
                \code{"solid"}), width, and color for standard deviation whiskers.}
          \item{sdstaplelty, sdstaplelwd, sdstaplecol}{Staple (end of whisker) line type,
                width, and color for standard deviation whiskers}
          \item{semwhisklty, semwhisklwd, semwhiskcol}{Whisker line type(default:
                \code{"solid"}), width, and color (default: \code{"#EDA217"})
                for standard error of the mean whiskers.}
          \item{semstaplelty, semstaplelwd, semstaplecol}{Staple (end of whisker) line type,
                width, and color (default: \code{"#EDA217"}) for standard error
                of the mean whiskers}
          \item{cistaplelty, cistaplelwd, cistaplecol}{Staple (end of whisker) line type,
                width, and color (default: \code{"#EDA217"}) for confidence interval
                whiskers}}}
}
\value{
Same as \code{\link[graphics]{boxplot}}
}
\description{
This is intended to be called from \code{\link{plotgroups}}.
}
\seealso{
\code{\link{plotgroups}}

\code{\link[graphics]{boxplot}}

Other plotgroups.plotting.functions: \code{\link{plotgroups.barplot}},
  \code{\link{plotgroups.beeswarm}},
  \code{\link{plotgroups.vioplot}}
}

