% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgroups.vioplot.R
\docType{data}
\name{plotgroups.vioplot}
\alias{plotgroups.vioplot}
\title{Plot several groups of repeated observations as a violin plot.}
\format{An object of class \code{list} of length 3.}
\usage{
plotgroups.vioplot
}
\arguments{
\item{data}{list, each element is a vector of replicates for one combination of parameters}

\item{at}{numeric vector giving the X locations where the plots should be drawn}

\item{stats}{list containing the summary statistics (see \code{\link{plotgroups}})}

\item{colors}{colors for plotting}

\item{features}{the features to plot.}

\item{barwidth}{width of the individual bars/boxes etc. as fraction of 1}

\item{boxpars}{parameters passed to \code{\link{plotgroups.boxplot}}}

\item{boxcol}{color of the boxes}

\item{boxwidth}{width of the boxes}

\item{...}{addtional parameters passed to \code{\link[vioplot]{vioplot}}}
}
\value{
List with the following components:
       \item{vioplot}{List containing the aggregated return values of \code{\link[vioplot]{vioplot}}}
       \item{boxplot}{Return value of \code{\link{plotgroups.boxplot}}}
}
\description{
This is intended to be called from \code{\link{plotgroups}}. Requires the \code{vioplot} package.
}
\seealso{
\code{\link{plotgroups}}

\code{\link[vioplot]{vioplot}}

Other plotgroups.plotting.functions: \code{\link{plotgroups.barplot}},
  \code{\link{plotgroups.beeswarm}},
  \code{\link{plotgroups.boxplot}}
}
\keyword{datasets}

