% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_inputs.R
\name{print_inputs}
\alias{print_inputs}
\title{Demonstrate ... Argument}
\usage{
print_inputs(...)
}
\arguments{
\item{...}{any collection of arguments}
}
\value{
nothing
}
\description{
This function just demonstrates how you can use the ... construct to pass any collection of arrays to a function for processing. Used as a prototype for the parallel_inputs function.
}
\examples{
print_inputs(1:3,1:5,1:9)
}
