% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_plot.R
\name{single_plot}
\alias{single_plot}
\title{Single R plot}
\usage{
single_plot()
}
\description{
Resets the plot window to only have a single frame. Many people's plotting code forgets to clean up after themselves.
}
