% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_all.R
\name{source_all}
\alias{source_all}
\title{Source all Scripts}
\usage{
source_all(directory, recursively = FALSE)
}
\arguments{
\item{directory}{string naming the folder to source}

\item{recursively}{logical, continue into subfolders too?}
}
\description{
This function will source every file within the target directory. It is not very useful but it is an interesting idea.
}
