% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title}
\alias{title}
\title{Plot Annotation}
\usage{
title(main = NULL, sub = NULL, xlab = NULL, ylab = NULL,
  main.line = NULL, sub.line = NULL, xlab.line = NULL, ylab.line = NULL,
  use.default.line = FALSE, ...)
}
\arguments{
\item{main}{Main plot title.}

\item{sub}{Plot subtitle.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{main.line}{The margin line for the main title. Defaults to \code{getOption('imisc.mgp.main')} or,
if the option is not set, 3.}

\item{sub.line}{The margin line for the sub-title. Defaults to \code{getOption('imisc.mgp.sub')} or,
if the option is not set, \code{par('mgp')[1] + 1}.}

\item{xlab.line}{The margin line for the X axis label. Defaults to \code{getOption('imisc.mgp.xlab')} or,
if the option is not set, \code{par('mgp')[1]}.}

\item{ylab.line}{The margin line for the Y axis label. Defaults to \code{getOption('imisc.mgp.xlab')} or,
if the option is not set, \code{par('mgp')[1]}.}

\item{use.default.line}{Use \code{\link[graphics]{title}} defaults instead of \code{getOption('imisc.mgp')},
even if the options are set.}

\item{...}{Additional parameters passed to \code{\link[graphics]{title}}}
}
\description{
This function can be used to add labels to a plot. Its main difference to \code{\link[graphics]{title}}
is that it allows to specify a different drawing position for each annotation. The defaults are the same
as for \code{\link[graphics]{title}}, but can be changed by a call to \code{\link[base]{options}}.
}
