% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgroups.R
\name{plotgroups}
\alias{plotgroups}
\title{Plot several groups of repeated observations, e.g. abundance/half-life of several
proteins each observed in several cell lines in several replicates. Observations can be grouped
either by protein (in which case cell lines will be annotated as X axis labels
and proteins above the plot) or by cell line.}
\usage{
plotgroups(data, names, colors = NULL, legend.text = NULL,
  legend.col = NULL, legend.pars = list(font = 2), names.split = NULL,
  names.italicize = NULL, names.style = c("plain", "combinatorial"),
  names.pch = 19, names.pch.cex = 1, names.pch.adj = 0.5,
  names.margin = 0.5, names.rotate = NULL, lwd.legend = NULL,
  features = c("median", "box", "iqr", "mean", "sd", "sem"), cex.xlab = 1,
  ylim = NULL, legendmargin = NULL, plot.fun = plotgroups.boxplot,
  plot.fun.pars = list(), barwidth = 0.8,
  ylab = deparse(substitute(data)), ...)
}
\arguments{
\item{data}{List, each element is a vector of replicates for one combination of parameters}

\item{names}{Character vector of X axis labels}

\item{colors}{Colors for plotting. Note that a group of observations is identified by consecutive
occurrence of the same color}

\item{legend.text}{Character vector of group names}

\item{legend.col}{Colors for group annotations. Defaults to plotting colors}

\item{legend.pars}{Parameters for group annotation. Will be passed to \code{\link[base]{text}}}

\item{names.split}{Character by which to split the \code{names}. Only useful in combination with
\code{names.italicize} or \code{names.style='combinatorial'}}

\item{names.italicize}{If a part of a \code{name} is to be written in italic text, the part is
identified by this character. I.e. The name is first split by \code{names.split}, each 
fragment containing \code{names.italicize} is rendered in italics}

\item{names.style}{How the \code{names} are to be rendered.
\describe{
      \item{plain}{Each name will be written as-is below the plot}
      \item{combinatorial}{Names will be split by \code{names.split}, unique strings will be
            printed at the bottom-left, and observations whose name contains the string will
            be identified by prining \code{names.pch} below the respective bar. Useful if e.g.
           assaying different combinations of single/double/triple knock-outs.}
           }}

\item{names.pch}{Character to be used for annotation of observations when
\code{names.style='combinatorial'}}

\item{names.pch.cex}{Character expansion factor for \code{names.pch}}

\item{names.pch.adj}{Text adjustment for \code{names.pch}. See \code{\link[base]{text}}}

\item{names.margin}{Spacing between the bottom edge of the plot and the annotation, in inches}

\item{names.rotate}{Only used when \code{names.style='plain'}. Degrees by which to rotate the
annotation strings.}

\item{lwd.legend}{Line width for grouping annotations. Defaults to \code{par("lwd")}}

\item{features}{Which features of the sample distributions to plot. Availability of features
depends on \code{plot.fun}}

\item{cex.xlab}{Character expansion factor for X axis annotation}

\item{ylim}{Y axis limits. Will be determined automatically if \code{NULL}}

\item{legendmargin}{Spacing between the upper-most data point/feature and the upper edge of the
plot, required for group annotation. Will be determined automatically if \code{NULL}}

\item{plot.fun}{Function to do the actual plotting. See \code{\link{plot.groups.boxplot}},
\code{\link{plot.groups.beeswarm}}}

\item{plot.fun.pars}{Additional parameters to pass to \code{plot.fun}}

\item{barwidth}{Width of the individual bars/boxes etc. as fraction of 1}

\item{ylab}{Y axis label}

\item{...}{Additional parameters passed to \code{\link[base]{par}}}
}
\description{
Plot several groups of repeated observations, e.g. abundance/half-life of several
proteins each observed in several cell lines in several replicates. Observations can be grouped
either by protein (in which case cell lines will be annotated as X axis labels
and proteins above the plot) or by cell line.
}
\examples{
data <- list()
for (i in 1:14) data[[i]] <- rnorm(50, i, 0.5)
names <- rep(c('gene1', 'gene2', 'gene3', 'gene1 gene2', 'gene1 gene3', 'gene2 gene3', 'gene1 gene2 gene3'), times=2)
colors <- rep(c("green", "blue"), each=7)
legend.text <- c("protein1", "protein2")
plotgroups(data, names, colors, legend.text, plot.fun=plotgroups.beeswarm, features=c('mean', 'sd'))
plotgroups(data, names, colors, legend.text, plot.fun=plotgroups.beeswarm, features=c('mean', 'sd'), names.style='combinatorial', names.split=" ", names.pch='Δ', plot.fun.pars=list(palpha=0.5, bxpcols="black"))
plotgroups(data, names, colors, legend.text, names.style='combinatorial', names.split=" ", names.pch='Δ')
}

